semester 4/Script Languages/sl-lab9-thu13-3/requirements.txt
# Add libraries required by your application
python-dotenv
openpyxl
csv